Flattened data directory, changed Markdown ext.
The initial plan was to have a search path for documentation that mirrored the
user's PATH. The leading slash would be removed from each directory in their
path, the resultant fragment appended to the path of the girl data directory,
then the girl client would search each directory in turn, returning after the
first match. I suspect that this is overkill, however, so for now I'll store
all the data in the same directory.

I've also changed the filename extension for the data files because Ikiwiki,
which I'll be using for the public wiki, requires the '.mdwn' extension.