Wrapped Girl in Module; created Pearl object.
The Girl class is now encompassed by a GirlDoc module so we'll be able to
throw our own exception objects in the future.

Pearls were previously represented as Strings, and Arrays thereof. They are
now Pearl objects, and Arrays of the same. This lets us abstract the file
handling and data directory code from the Girl class.