Rudimentary rendering class and pager logic.
Added provisional support for conditional paging of output. If the pearl will
fit on one screen it's printed to STDOUT; else, it's piped to the user's
pager. There are no tests yet for this feature, mainly because I'm unsure how
to confirm what has been piped to a pager from a test script.