Added ANSI formatter, fixed pager selection bug.
Created formatter that performs selective colour highlighting using ANSI. It's
imperfect, though, because it does not distinguish between pagers/terminals
that support colour, and those that can't. For instance, `less` just shows
garbled nonsense.

In addition, the code that chose which pager to use was buggy: it only ever
considered the value of ENV['PAGER']; ignoring the PAGERS constant. This has
been fixed.