Don't use ANSI colour escapes when page is `less`
`less` doesn't understand ANSI colour escapes, so when presented them it
garbles the output. We were previously unable to format the output especially
for `less` because we didn't know the pager name; we simply acted on an IO
object that had already been created with the pager command. It was difficult
to get it before the page{} block, because, if we were to choose a pager that
IO.popen was willing to open a pipe to, then we had to create the IO object
before we knew whether the pager being considered would work. However, by
creating the IO object we screwed up STDOUT... Anyway, we now save each pager
name that we consider to an instance variable, so the name that variable
contains after we are in the page{} block is the pager being used. To take
advantage of this, we must determine which formatter to use while inside the
page{} block. This isn't ideal, but works for now.

There was a further problem, that's also been solved now. After checking the
PAGER environment variable, we look for a pager named `pager`. This is
typically a symlink to the user's default pager. However, on Debian systems, at
least, `pager` is a symlink to `\etc\alternatives\pager`, which is itself a
symlink to the actual pager binary. The reason that this matters is that we
need to know whether `pager` points to `less`, because if so we need to format
the output in plain text, and to do so we need to (a) find the real path of
`pager`, (b) resolve the symlinks in that path, and (c) return the pager name.
We can't, directly, use File.readlink for part b of this task because it
doesn't work recursively... To solve this problem we have a File.whereis method
which searches the users PATH for a command and returns the first existing
pathname found, and a .pager_name method which uses File.basename and
Pathname's .realpath to handle parts a and c.